requirements.txt
flask
celery
openai-whisper
psycopg2-binary
python-dotenv
SpeechRecognition
pyaudio
